refactor: separate session spawn context ownership
